sparv-sbx-sentence-sentiment-kb-sent/pyproject.toml
[project]
name = "sparv-sbx-sentence-sentiment-kb-sent"
version = "0.1.0"
description = "A sparv plugin for computing word neighbours using a BERT model."
authors = [
    { name = "Språkbanken Text", email = "sb-info@svenska.gu.se" },
    { name = "Kristoffer Andersson", email = "kristoffer.andersson@gu.se" },
]
dependencies = ["sparv-pipeline >=5.2.0", "transformers>=4.34.1"]
license = "MIT"
readme = "README.md"
requires-python = ">= 3.8.1,<3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    # "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    # "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    # "uncomment if you test on these interpreters:",
    # "Programming Language :: Python :: Implementation :: IronPython",
    # "Programming Language :: Python :: Implementation :: Jython",
    # "Programming Language :: Python :: Implementation :: Stackless",
    "Topic :: Utilities",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."sparv.plugin"]
sbx_sentence_sentiment_kb_sent = "sbx_sentence_sentiment_kb_sent"

[project.urls]
Homepage = "https://spraakbanken.gu.se"
Repository = "https://github.com/spraakbanken/sparv-sbx-sentence-sentiment-analysis"
"Bug Tracker" = "https://github.com/spraakbanken/sparv-sbx-sentence-sentiment-analysis/labels/project%3Asentence-sentiment--kb-sent"

[tool.hatch.build.targets.sdist]
exclude = ["/.github", "/docs"]

[tool.hatch.build.targets.wheel]
packages = ["src/sbx_sentence_sentiment_kb_sent"]

[tool.hatch.metadata]
allow-direct-references = true


[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.0.0",
    "syrupy>=4.6.1",
]
